requirements.txt
httpx==0.24.1
python-dotenv